requirements.txt
mistune==3.0.2
langchain_chroma==0.1.4
pydantic==2.10.1
chainlit==1.3.2
transformers==4.47.0
torch==2.5.1
unstructured==0.16.10
docx2txt==0.8
docxcompose==1.4.0
bitsandbytes==0.45.0
chromadb==0.5.23
exception==0.1.0